release: 0.24.0 USB fractional speeds, check exit codes for scripting
- USBDevice.SpeedMbps is float64: sysfs reports low-speed devices as
  '1.5' which the previous int64 truncated to 1 Mb/s; SpeedString()
  renders whole rates without decimals and keeps '1.5 Mb/s' (JSON stays
  backward-compatible with old integer captures).
- check now ends with a Result: summary line and exits non-zero:
  1 warning findings, 2 critical findings, 3 collector errors
  (precedence 3 > 2 > 1 > 0 since an incomplete capture cannot be
  trusted). --no-fail opts out for pipelines that only need output.
  Exit-code table documented in the command help and USERS_MANUAL.

+++ internal/collect/usb_power.go
@@ -49,7 +49,7 @@ func readUSBDevices(dir string) []model.USBDevice {
 		device.Class = readSysString(base, "bDeviceClass")
 		if speed := readSysString(base, "speed"); speed != "" {
 			if mbps, parseErr := strconv.ParseFloat(speed, 64); parseErr == nil {
-				device.SpeedMbps = int64(mbps)
+				device.SpeedMbps = mbps
 			}
 		}
 		if device.VendorID == "" && device.ProductID == "" && device.Product == "" {
